update submodules for ADI fw v0.37
This updates the submodules to sync up with ADI plutosdr-fw release
v0.37.

buildroot and linux have been moved to the branches that ADI is using
now, with the maia-sdr commits being cherry-picked on top of the new
branches.

maia-sdr has been updated to the latest main, which includes the correct
version of the adi-hdl submodule.

Signed-off-by: Daniel Estévez <daniel@destevez.net>